Add text input class modifier
This commit adds the ability to render the appropriate sized input
box for the allowed size of the field.

Fields limited to less that 20 characters will render as a 25% field,
Fields between 20 and 59 characters will render a 50% field, and fields
between 60 and 499 characters will render a 75% text box.

Fields which have no maxLength attribute attached will render at the
default 100% width.