Add answer helper
* Add auto-complete to birthday fields

This commit will add the auto-complete attribute to birthday fields, in compliance
with the govUk design standard.

This commit also adds functionality to change the default "continue" text on a
form button.

* Add select transformer

This commit adds a select transformer. This renders as a drop down menu
on a govuk webpage.

The transformer will now render a select element if the schema is type
string, and has a oneOf array of length greater than or equal to 20.

* Add text input class modifier

This commit adds the ability to render the appropriate sized input
box for the allowed size of the field.

Fields limited to less that 20 characters will render as a 25% field,
Fields between 20 and 59 characters will render a 50% field, and fields
between 60 and 499 characters will render a 75% text box.

Fields which have no maxLength attribute attached will render at the
default 100% width.

* Removed hard-coded class assignment

* Fix bug in unit tests - expect updated to reflect changes

* Minor changes suggested in pull request

* Move 'auto complete' and 'is summary' to the uischema

Previously the implementation of the 'auto-complete' option for date fields, and
distinguishing a summary page from a standard question was handled in the Schema

As this is presentation logic, it has now been moved into the UISchema.

* change irrelevant comment

* Add summary transformer

This commit will add a summary transformer to build a
'check your answers' page.

* Refactor answer helper

The answer helper was previously part of the cica-web module. However,
since the creation of the summary transformer, it makes more sense
to include the answer helper as part of this module. This means answers
are only formatted when intended, and also removes undesirable logic
from the cica-web module.

* Fix date handling

This commit will implement changes to handling partial dates. These
will now be displaying correctly, and pre-populated fields will be
accurately populated.

* refactor 'default' transformer for type integer && reword police lookup test